-observer design pattern implemented. -App runs better now.

+++ app/src/main/java/com/example/orderyurt/Discount/Coupon.java
@@ -5,12 +5,12 @@ import java.util.Date;
 public class Coupon {
     private String title;
     private String code; //CouponID in DB
-    private float value;
+    private Double value;
     private Date startDate;
     private Date endDate;
     private int RID;
 
-    public Coupon(String title, String code, float value, Date startDate, Date endDate, int RID){
+    public Coupon(String title, String code, Double value, Date startDate, Date endDate, int RID){
         this.title = title;
         this.code = code;
         this.value = value;
